Prep for merge into main
Grab out the demo into a specific file, restore app.dart to have barebones struct, so that other demos/the real flow can be loaded in the build function.
